Corrected answers to ch1/pg26 problems
I skipped two 2hole problems, and redid them

I'm pretty sure that my answer in ch1 pg26 #7 is the same as the answer in the Answers section, but I'm not 100% sure.

Also, it appears sometimes that in the solutions, like in #1 at the step

(\a.\b.\c.cba)(z)z(\w.\v.w)
(\b.\c.\cbz)(z)(\w.\v.w)

it looks like the right-most z was applied first, when I would expect the left-most z, in the parenthesis, to be applied first.
I think this is just the way the parenthesis are switched around.